Reorganise deploy scripts
