+++ scripts/deploy/protocol/02_deployScoutProtocolBuilderNft.ts
